+++ src/main/java/app/Twiter/controllers/UserController.java
@@ -1,18 +1,15 @@
-package app.Twiter.controllers.user_inter;
+package app.Twiter.controllers;
 
 import app.Twiter.model.User;
 import app.Twiter.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-public class SearchController {
+public class UserController {
 
     @Autowired
     private UserService userService;
@@ -26,5 +23,9 @@ public class SearchController {
         return userService.getByID(id);
     }
 
+    @PostMapping(value = "/users", consumes = MediaType.APPLICATION_JSON_VALUE)
+    public void registerUser(@RequestBody User user){
+        userService.registerUser(user);
+    }
 
 }
