--- src/main/java/app/Twiter/controllers/like/LikeController.java
@@ -1,9 +1,0 @@
-package app.Twiter.controllers.like;
-
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class LikeController {
-
-}
