--- src/main/java/app/Twiter/controllers/post/DeletePostController.java
@@ -1,4 +1,0 @@
-package app.Twiter.controllers.post;
-
-public class DeletePostController {
-}
