--- src/main/java/app/Twiter/controllers/reply/AddPostReplyController.java
@@ -1,8 +1,0 @@
-package app.Twiter.controllers.reply;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class AddPostReplyController {
-
-}
