--- src/main/java/app/Twiter/controllers/user_inter/FollowController.java
@@ -1,13 +1,0 @@
-package app.Twiter.controllers.user_inter;
-
-import app.Twiter.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class FollowController {
-
-    @Autowired
-    UserService userService;
-}
