--- src/main/java/app/Twiter/controllers/user_inter/RegisterController.java
@@ -1,23 +1,0 @@
-package app.Twiter.controllers.user_inter;
-
-import app.Twiter.model.User;
-import app.Twiter.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.awt.*;
-
-@RestController
-public class RegisterController {
-
-    @Autowired
-    private UserService userService;
-
-    @PostMapping(value = "/users", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public void registerUser(@RequestBody User user){
-        userService.registerUser(user);
-    }
-}
