--- src/main/java/app/Twiter/controllers/user_inter/UnfollowController.java
@@ -1,4 +1,0 @@
-package app.Twiter.controllers.user_inter;
-
-public class UnfollowController {
-}
