--- src/main/java/app/Twiter/controllers/user_inter/UnregisterController.java
@@ -1,4 +1,0 @@
-package app.Twiter.controllers.user_inter;
-
-public class UnregisterController {
-}
